refactor: rename base
