+++ truelearn/learning/_engage_classifier.py
@@ -2,7 +2,7 @@ from typing import Dict, Any
 from typing_extensions import Self
 
 from truelearn.models import EventModel
-from ._base import BaseClassifier
+from truelearn.base import BaseClassifier
 
 
 class EngageClassifier(BaseClassifier):
