+++ truelearn/learning/_majority_classifier.py
@@ -2,7 +2,7 @@ from typing import Any, Dict
 from typing_extensions import Self
 
 from truelearn.models import EventModel
-from ._base import BaseClassifier
+from truelearn.base import BaseClassifier
 from ._constraint import TypeConstraint
